+++ src/main/java/org/team6/coffeebeanery/order/model/Order.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer orderId; //주문 ID
+    private Long orderId; //주문 ID
     
     @Column(length = 254)
     private String customerEmail; //고객 이메일
@@ -26,7 +26,7 @@ public class Order {
     
     private LocalDateTime orderCreatedAt; //주문 날짜 및 시간
     
-    private Double totalPrice; //총 주문 금액
+    private Long totalPrice; //총 주문 금액
     
     @Enumerated(EnumType.STRING)
     private OrderStatus orderStatus; //주문 상태
